Vector jednak będzie w double bo sinusa będę używał

+++ include/utils.h
@@ -6,13 +6,13 @@
 //      ale myślę że wystarczy dać tu klika dobrych funkcji i będzie spokój
 typedef struct _vector {
         int size;
-        int* value;
+        double* value;
 } Vector;
 
 Vector* Vector_Construct(int s);
-void Vector_Append(Vector* v, int value);       
+void Vector_Append(Vector* v, double value);       
 int Vector_ChangeSize(Vector* v, int s);        //zwraca poprzednią długość
-int Vector_Pop(Vector* v);                      //zwraca usuniętą wartość
-int Vector_Delete(Vector* v, int index);        //zwraca usuniętą wartość
+double Vector_Pop(Vector* v);                      //zwraca usuniętą wartość
+double Vector_Delete(Vector* v, int index);        //zwraca usuniętą wartość
 
 #endif
