+++ src/utils.c
@@ -4,16 +4,16 @@
 Vector* Vector_Construct(int s) {
         Vector* temp = malloc(sizeof(Vector));
         temp->size = s;
-        temp->value = malloc(s * sizeof(int));
+        temp->value = malloc(s * sizeof(double));
 
         return temp;
 }
 
-void Vector_Append(Vector* v, int value) {
+void Vector_Append(Vector* v, double value) {
         if (v == NULL) return;
 
         v->size++;
-        v->value = realloc(v->value, v->size * sizeof(int));
+        v->value = realloc(v->value, v->size * sizeof(double));
         v->value[v->size - 1] = value;
 }
 
@@ -21,30 +21,30 @@ int Vector_ChangeSize(Vector* v, int s) {
         if (v == NULL) return -1;
 
         int temp = v->size;
-        v->value = realloc(v->value, s * sizeof(int));
+        v->value = realloc(v->value, s * sizeof(double));
         v->size = s;
         return temp;
 }
 
-int Vector_Pop(Vector* v) {
+double Vector_Pop(Vector* v) {
         if (v == NULL) return 1;
 
         if (v->size > 0) {
-                int temp = v->value[v->size - 1];
+                double temp = v->value[v->size - 1];
 
                 v->size--;
-                v->value = realloc(v->value, v->size * sizeof(int));
+                v->value = realloc(v->value, v->size * sizeof(double));
                 
                 return temp;
         } else return 1;
 }
 
-int Vector_Delete(Vector* v, int index) {
+double Vector_Delete(Vector* v, int index) {
         if (v == NULL) return 1;
 
         v->size--;
-        int temp = v->value[index] = v->value[v->size];
-        v->value = realloc(v->value, v->size * sizeof(int));
+        double temp = v->value[index] = v->value[v->size];
+        v->value = realloc(v->value, v->size * sizeof(double));
 
         return temp;
 }
